user access blocked of updating other users

+++ app/controllers/user.controller.js
@@ -1,18 +1,9 @@
 const router = require('express').Router();
 const User = require('../models/user.model')
-
-// Create a new user
-// router.post('/', async (req, res) => {
-//   try {
-//     const createdUser = await User.create(req.body);
-//     res.status(201).json(createdUser);
-//   } catch (error) {
-//     res.status(500).json({ error: 'Failed to create user!!! ' + error.message });
-//   }
-// });
+const adminMiddleware = require("../middleware/admin.middleware")
 
 // Get all users
-router.get('/', async (req, res) => {
+router.get('/', adminMiddleware, async (req, res) => {
   try {
     const users = await User.find({ ...req.query });
     res.json(users);
@@ -25,6 +16,8 @@ router.get('/', async (req, res) => {
 router.get('/:id', async (req, res) => {
   try {
     const id = req.params.id;
+    if (req.user.userType === "USER" && id !== req.user.id) return res.status(403).json({ message: "Don't have access to get other user details" })
+
     const user = await User.findById(id);
     if (!user) {
       return res.status(404).json({ error: 'User not found' });
@@ -39,6 +32,7 @@ router.get('/:id', async (req, res) => {
 router.patch('/:id', async (req, res) => {
   try {
     const id = req.params.id;
+    if (req.user.userType === "USER" && id !== req.user.id) return res.status(403).json({ message: "Don't have access to update other user details" })
 
     // `new: true` The method returns the updated document after applying the update operation.
     const updatedUser = await User.findOneAndUpdate({ _id: id }, { $set: req.body }, { new: true, runValidators: true })
@@ -56,6 +50,8 @@ router.patch('/:id', async (req, res) => {
 router.delete('/:id', async (req, res) => {
   try {
     const id = req.params.id;
+    if (req.user.userType === "USER" && id !== req.user.id) return res.status(403).json({ message: "Don't have access to delete other user details" })
+
     const deletedUser = await User.findByIdAndDelete(id);
     if (!deletedUser) {
       return res.status(404).json({ error: 'User not found' });
@@ -69,8 +65,10 @@ router.delete('/:id', async (req, res) => {
 // Get User by its username 
 router.get('/username/:id', async (req, res) => {
   try {
-    const id = req.params.id;
-    const user = await User.findOne({ username: id });
+    const username = req.params.id;
+    if (req.user.userType === "USER" && username !== req.user.username) return res.status(403).json({ message: "Don't have access to get other username details" })
+
+    const user = await User.findOne({ username });
 
     if (!user) {
       return res.status(404).json({ error: 'User does not exists with this Username' });
